requirements.txt
certifi==2019.3.9
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
djangorestframework==3.9.2
gunicorn==19.9.0
pipenv==2018.11.26
pytz==2019.1
sqlparse==0.3.0
virtualenv==16.5.0
virtualenv-clone==0.5.3
whitenoise==4.1.2

appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.1.2
enum34==1.1.6
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
psycopg2-binary